refactor(clan): unify validateTag & remove magic numbers
api:
- migrate all uses of validateClantag → clans/validateTag
- clan.entity & CreateClanDto use config.clanLength
- auth.change-clantag now wraps ClansService.join (legacy compat)

client:
- thunk changeClanAsync keeps original case
- prompt text updated to 1-7 chars

helpers/validateClantag now re-exports validateTag for fallback
