+++ client/src/ui/App.tsx
@@ -359,22 +359,13 @@ function App() {
   };
   const onChangeClan = () => {
     const newClan = prompt(
-      'What do you want your clan tag to be? Clans can only be 1-4 characters long, and you can only change your clan once every 7 days.',
+      'What do you want your clan tag to be? Clans can only be 1-7 characters long, and you can only change your clan once every 7 days.',
     );
     if (!newClan) return;
 
     dispatch(changeClanAsync(newClan) as any);
   };
-  /*
 
-  Doesn't work find an alternative
-
-  const onRemoveClan = () => {
-    const newClan = prompt('Are you sure you want to remove your clan tag? This can only be done once every 7 days. Type anything to confirm, or press "cancel" to exit.');
-
-    dispatch(changeClanAsync('') as any);
-  }
-    */
   const openShop = () => {
     setModal(<ShopModal account={account} />);
   };
